requirements.txt
# =============================================================================
# MedSwarm Dependencies
# =============================================================================

# Geographic & Graph Processing
osmnx>=1.3.0
networkx>=3.0

# Math & Data
numpy>=1.24.0
pandas>=2.0.0

# Reinforcement Learning
gymnasium>=0.28.1
stable-baselines3>=2.0.0

# Visualization & Dashboard
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
plotly>=5.18.0

# Configuration
pyyaml>=6.0.0

# Progress bars
tqdm>=4.65.0